add projectile pooling system and splash damage #348 #217 #417 (#392)
* add projectile pooling system and splash damage #348 #217
* Add parabolic arc projectile with AoE landing damage #411
* add multi-shot system for towers #412
* Fix defensive building not shooting after restoration from ruins #417
* refactor damage system to use integers and improve projectile pooling #348

+++ Lords_Frontiers/Source/Lords_Frontiers/Public/Units/Unit.h
@@ -45,7 +45,7 @@ public:
 
 	virtual void Attack( TObjectPtr<AActor> hitActor ) override;
 
-	virtual void TakeDamage( float damage ) override;
+	virtual void TakeDamage( int damage ) override;
 
 	void AdvancePathPointIndex();
